Introduced NetworkUtil for serialization of data to send/receive between client and server; it's now used in the server, not yet in the client. The bug of the overflowing track position has been fixed.

+++ common/serverconnection.cpp
@@ -431,25 +431,27 @@ namespace PMP {
             quint8 volume = message[3];
 
             quint32 queueLength =
-                (message[4] << 24)
-                + (message[5] << 16)
-                + (message[6] << 8)
-                + message[7];
+                ((quint8)message[4] << 24)
+                + ((quint8)message[5] << 16)
+                + ((quint8)message[6] << 8)
+                + (quint8)message[7];
 
             quint32 queueID =
-                (message[8] << 24)
-                + (message[9] << 16)
-                + (message[10] << 8)
-                + message[11];
+                ((quint8)message[8] << 24)
+                + ((quint8)message[9] << 16)
+                + ((quint8)message[10] << 8)
+                + (quint8)message[11];
 
-            quint64 position = message[12];
-            position = (position << 8) + message[13];
-            position = (position << 8) + message[14];
-            position = (position << 8) + message[15];
-            position = (position << 8) + message[16];
-            position = (position << 8) + message[17];
-            position = (position << 8) + message[18];
-            position = (position << 8) + message[19];
+            //qDebug() << "received position bytes" << (quint8)message[12] << (quint8)message[13] << (quint8)message[14] << (quint8)message[15] << (quint8)message[16] << (quint8)message[17] << (quint8)message[18] << (quint8)message[19];
+
+            quint64 position = (quint8)message[12];
+            position = (position << 8) + (quint8)message[13];
+            position = (position << 8) + (quint8)message[14];
+            position = (position << 8) + (quint8)message[15];
+            position = (position << 8) + (quint8)message[16];
+            position = (position << 8) + (quint8)message[17];
+            position = (position << 8) + (quint8)message[18];
+            position = (position << 8) + (quint8)message[19];
 
             qDebug() << "track position:" << position;
